[AIP-21] Move contrib.sensors.weekday_sensor sensors.weekday_sensor
